move `replaceRaw` to `replace`
