--- src/lib/helpers/replaceRaw.js
@@ -1,58 +1,0 @@
-const util = require('util')
-const dotenv = require('dotenv')
-
-const escapeForRegex = require('./escapeForRegex')
-
-function replaceRaw (src, key, replaceValue) {
-  let output
-  let escapedValue = util.inspect(replaceValue, { showHidden: false, depth: null, colors: false })
-  if (replaceValue.includes('\n')) {
-    escapedValue = JSON.stringify(replaceValue) // use JSON stringify if string contains newlines
-    escapedValue = escapedValue.replace(/\\n/g, '\n') // fix up newlines
-    escapedValue = escapedValue.replace(/\\r/g, '\r')
-  }
-  let newPart = `${key}=${escapedValue}`
-
-  const parsed = dotenv.parse(src)
-  if (Object.prototype.hasOwnProperty.call(parsed, key)) {
-    const originalValue = parsed[key]
-    const escapedOriginalValue = escapeForRegex(originalValue)
-
-    // conditionally enforce end of line
-    let enforceEndOfLine = ''
-    if (escapedOriginalValue === '') {
-      enforceEndOfLine = '$' // EMPTY scenario
-    }
-
-    const currentPart = new RegExp(
-      '^' + // start of line
-      '(\\s*)?' + // spaces
-      '(export\\s+)?' + // export
-      key + // KEY
-      '\\s*=\\s*' + // spaces (KEY = value)
-      '["\'`]?' + // open quote
-      escapedOriginalValue + // escaped value
-      '["\'`]?' + // close quote
-      enforceEndOfLine
-      ,
-      'gm' // (g)lobal (m)ultiline
-    )
-
-    // $1 preserves spaces
-    // $2 preserves export
-    output = src.replace(currentPart, `$1$2${newPart}`)
-  } else {
-    // append
-    if (src.endsWith('\n')) {
-      newPart = newPart + '\n'
-    } else {
-      newPart = '\n' + newPart
-    }
-
-    output = src + newPart
-  }
-
-  return output
-}
-
-module.exports = replaceRaw
