+++ tests/monorepo/apps/backend/.env.example
@@ -1,2 +1,2 @@
 # .env.example - generated with dotenvx
-HELLO=""
+HELLO=''
